backend/requirements.txt
django
django-rest-framework
django-cors-headers
psycopg2-binary
dj-database-url